# app/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/messages": {
            "post": {
                "description": "Create a new message with content and recipient.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "messages"
                ],
                "summary": "Create a new message",
                "parameters": [
                    {
                        "description": "Message request payload",
                        "name": "message",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.AddMessageRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success response",
                        "schema": {
                            "$ref": "#/definitions/models.AddMessageResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request response",
                        "schema": {
                            "$ref": "#/definitions/models.CreateMessageErrorResponse"
                        }
                    }
                }
            }
        },
        "/messages/processing/start": {
            "post": {
                "description": "Start the message sending job",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "processing"
                ],
                "summary": "Start message processing",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.MessageProcessingResponse"
                        }
                    }
                }
            }
        },
        "/messages/processing/stop": {
            "post": {
                "description": "Stop the message sending job",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "processing"
                ],
                "summary": "Stop message processing",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.MessageProcessingResponse"
                        }
                    }
                }
            }
        },
        "/messages/sent": {
            "get": {
                "description": "Retrieve all sent messages",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "messages"
                ],
                "summary": "Get sent messages",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.SentMessageResponseModel"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.AddMessageRequest": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "recipient": {
                    "type": "string"
                }
            }
        },
        "models.AddMessageResponse": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "recipient": {
                    "type": "string"
                }
            }
        },
        "models.CreateMessageErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "models.MessageProcessingResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "models.SentMessageResponseModel": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "recipient": {
                    "type": "string"
                }
            }
        }
    }
}